backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-ratelimit==4.1.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
dj-database-url==2.2.0
boto3==1.34.131
django-storages==1.14.3
Faker==25.8.0
python-decouple==3.8
Pillow>=11.0.0
gunicorn==22.0.0